+++ tests/Endpoint/ResourceEndpointTest.php
@@ -7,6 +7,8 @@ namespace DigitalCz\DigiSign\Endpoint;
 use DigitalCz\DigiSign\DigiSignClient;
 use DigitalCz\DigiSign\Exception\EmptyResultException;
 use DigitalCz\DigiSign\Exception\ResponseException;
+use DigitalCz\DigiSign\Resource\BaseResourceTest;
+use DigitalCz\DigiSign\Resource\Collection;
 use DigitalCz\DigiSign\Resource\DummyResource;
 use Nyholm\Psr7\Response;
 use PHPUnit\Framework\TestCase;
@@ -177,6 +179,21 @@ class ResourceEndpointTest extends TestCase
         $endpoint = new DummyEndpoint($parent);
         $resource = $endpoint->list(['foo' => 'bar']);
         self::assertSame(DummyResource::LIST_EXAMPLE, $resource->getResult());
+    }
+
+    public function testCreateCollectionResource(): void
+    {
+        $expectedResponse = new Response(200, [], DigiSignClient::jsonEncode(DummyResource::COLLECTION_EXAMPLE));
+
+        $parent = $this->createMock(EndpointInterface::class);
+        $endpoint = new DummyEndpoint($parent);
+        /** @var Collection<DummyResource> $resource */
+        $resource = $endpoint->collection($expectedResponse);
+        self::assertSame([
+            BaseResourceTest::DUMMY_RESOURCE_ARRAYED,
+            BaseResourceTest::DUMMY_RESOURCE_ARRAYED,
+            BaseResourceTest::DUMMY_RESOURCE_ARRAYED,
+        ], $resource->toArray());
     }
 
     public function testResponseExceptionEmptyBody(): void